Fix bug in search when scrolling with arrow keys
When the mouse hovered over a search result while the results are being scrolled using the arrow keys, the selection would jump back to the hovered element every time the selection reaches the end of the window. To fix this, prevent hovering while using the arrow keys and reenable hovering when the mouse is actually moved.