ConceptTree: Prevent conflicts in chooseFromUri
This is related to issue #16. As chooseFromUri is a complicated action which not only requires waiting for API requests, but all kinds of different operations, cancel tokens were not used. Instead, for each chooseFromUri call a unique identifier is generated and saved and then at each step this identifier is compared with the saved one. If it changed in the meantime, the request is canceled. While testing, this proved to work very well and consistent.