Allow creation of mappings with no target concepts
This needed several internal changes and might have changed other behavior as well. For example, When loading the application using URL parameters and a mapping identifier exists, the encoded mapping will still be decoded and used in MappingEditor. This was necessary because some actions inside of the application caused a reload from URL parameters which caused changes in MappingEditor to be overridden.