Vuex settings module: Possible fix for rare issue
Seemingly at random, the creator (and possibly other settings) were reset after reloading the page. It might have been a race condition with loading the settings from local storage which is an asynchronous task. So to fix this, it is not possible anymore to modify the settings before they were loaded from local storage.