AnnotationPopover: Add workaround to mitigate slightly buggy behavior
When showing the annotations for a mapping, if the mouse was moved direcly to the next mapping, the content of the AnnotationPopover was adjusted accordingly, but the position was not moved to the next mapping. In order to mitigate this, the popover is hidden and then shown again if the mapping changes.

Unfortunately, this doesn't work perfectly either. Sometimes, when performing the same movement, the AnnotationPopover won't show until the mouse is moved again, even if only a pixel. But this result is still better than the previous behavior.